feat: Podman icon
Adds a Podman icon (woff2 font with only one character, the seal) and
uses it to set a custom icon for any container with the 'io.podman.demo'
label.

For now this is just an easter egg, but it's a start to using the podman
icon elsewhere in the future. (podman-pause infra containers? status-bar
actions?)

Fixes #3307.

Signed-off-by: Tim deBoer <git@tdeboer.ca>